restructure directory: - Change name to wayback_google_analytics - Update imports to reflect this
